requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
openai==1.13.3
packaging==23.2
pydantic==2.6.1
pydantic-core==2.16.2
python-dotenv==1.0.1
sniffio==1.3.0
termcolor==2.4.0
tqdm==4.66.2
typing-extensions==4.9.0
